Treat character int literals with non-word chars as numbers
Before, a literal like `0'.` would be highlighted as the number `0` and
the unterminated string beginning `'.`. This was due to the `\b` word
boundary assertion at the end of the `number` pattern failing to match
after non-word characters.

Now, the pattern does not assert a word boundary at the end for that
flavor of number, only at the beginning. It now correctly highlights
the problematic example.

Fixes #1.